requirements.txt
Flask~=2.2.5
flask_login~=0.6.3
SQLAlchemy~=2.0.23
WTForms~=3.0.1
Werkzeug~=2.2.3
importlib-metadata==4.13.0
pymysql~=1.1.0
Flask-WTF~=1.1.1
email_validator~=2.0.0
celery~=5.2.7
redis~=4.4.0
stripe~=7.12.0
CORS==1.0.1
flask-cors~=4.0.0
azure-storage-blob~=12.19.0
scikit-learn~=1.0.2
pandas~=1.3.5
numpy~=1.21.6
